requirements.txt
streamlit==1.21.0
numpy==1.24.3
folium==0.14.0
matplotlib-inline==0.1.6
haversine==2.8.0
streamlit-folium==0.11.1
Pillow==9.5.0